+++ apps/mobile/src/app/(home,search,profile,notifications,settings)/_layout.tsx
@@ -262,6 +262,9 @@ function StackLayout({ children }: PropsWithChildren) {
           },
         })}
         name="posts/[id]/index"
+        options={{
+          headerTransparent: false,
+        }}
       />
 
       <Stack.Screen